fix(tests): fix remaining compile errors and NaN bit pattern injection
- Fix Pattern B in test_accuracy.mojo:164 (Float32(Float32()) with value on next line)
- Fix unused variable 'layer_sizes' in test_core_operations.mojo (rename to _layer_sizes)
- Fix unused variable 'loss_tracker' in test_metrics_coordination.mojo (rename to _loss_tracker)
- Fix set(UInt32/UInt64) in AnyTensor to bitcast for float32/float64 tensors, preserving
  NaN bit patterns — previously silently dropped writes on float dtypes via _set_int64
- Fix ASAN workflow: replace dtype tests (incompatible with ASAN JIT) with hash tests

+++ tests/training/test_accuracy.mojo
@@ -161,9 +161,7 @@ fn test_topk_accuracy_k1() raises:
             if c == true_class:
                 logits.set(idx, Float32(10.0))
             else:
-                logits.set(idx, Float32(Float32())
-                    c
-                )  # Lower scores
+                logits.set(idx, Float32(c))  # Lower scores
 
     var acc_top1 = top1_accuracy(logits, labels)
     var acc_topk = topk_accuracy(logits, labels, k=1)
